requirements.txt
# requirements.txt - UPDATED FOR CURRENT PYTORCH VERSIONS
torch>=2.2.0,<2.9.0
torchvision>=0.17.0,<0.19.0
transformers>=4.35.0
captum>=0.7.0
gradio>=4.19.0
Pillow>=10.0.0
matplotlib>=3.7.0
numpy>=1.24.0
requests>=2.28.0